+++ R/misc.R
@@ -15,6 +15,7 @@ NULL
 #' @importFrom magick image_canny
 #' @importFrom magick image_charcoal
 #' @importFrom magick image_composite
+#' @importFrom magick image_convert
 #' @importFrom magick image_crop
 #' @importFrom magick image_despeckle
 #' @importFrom magick image_emboss
@@ -25,6 +26,7 @@ NULL
 #' @importFrom magick image_oilpaint
 #' @importFrom magick image_quantize
 #' @importFrom magick image_reducenoise
+#' @importFrom magick image_scale
 #' @importFrom magick image_threshold
 #' @importFrom magick image_write
 #' @importFrom magick magick_config